+++ CHANGES.md
@@ -15,7 +15,8 @@
 ### UPDATE
 
 - Kotlin を 1.2.60 に上げた
-
+- CircleCI のコンテナに入っている Android NDK を使うよう変更した
+- ビルドメモリ設定を `gradle.properties` から `JVM_OPTS: -Xmx3200m` に変更した
 
 ## 1.7.0