+++ CHANGES.md
@@ -14,9 +14,10 @@
 
 ### UPDATE
 
-- Kotlin を 1.2.60 に上げた
+- Kotlin を 1.2.71 に上げた
 - CircleCI のコンテナに入っている Android NDK を使うよう変更した
 - ビルドメモリ設定を `gradle.properties` から `JVM_OPTS: -Xmx3200m` に変更した
+- Android Studio 3.2.0 に対応した
 
 ## 1.7.0